Update serverless-rest-api-assignment-stack.ts
This commit adds functionality to deploy a static website to an S3 bucket. It creates a new S3 bucket named "SiteBucket" with specified properties such as public read access, removal policy, and website index document. Additionally, it utilizes the aws-cdk-lib/aws-s3-deployment module to deploy website assets from the "./dist" directory to the newly created bucket. Finally, it outputs the URL of the deployed website as a CloudFormation output.